Fix broken waterfall after commit of RemovePaddingTransform.
In r1226 PERelinker was modified to employ RemovePaddingTransform in its pipeline. This caused problems with order generation as Playback uses the original decomposition (including padding blocks), which is then fed to Reorderer. Thus, padding blocks make it in to an order specification and ExplicitOrderer eventually fails when invoked by PERelinker as the padding blocks no longer exist at this time.

This CL addresses the issue by introducing CopyImageLayoutWithoutPadding to image_layout.h/.cc, which is invoked by both Playback and PERelinker immediately after decomposition. This ensures that both halves of the workflow use the same decomposition as a starting point.

This makes RemovePaddingTransform irrelevant, thus it has been removed.

A new test_data target and relink.exe unittest was added to tickle this bug.

BUG=

Review URL: https://codereview.appspot.com/6854045

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@1227 15e8cca8-e42c-11de-a347-f34a4f72eb7d