Create filter executable.
The filter executable is used for creating an address-space filter, to be used for selectively instrumenting a binary. This implements the shell of the executable, and the simple set operations. The meat of the program (compiling a filter from a text description and symbols) will come in later CLs.

BUG=

Review URL: https://codereview.appspot.com/7654043

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@1382 15e8cca8-e42c-11de-a347-f34a4f72eb7d