Create SampledModuleCache.
This CL creates the workhorse of the collection side of the sampling profiler. This takes care of managing handles to processes and modules, as well as managing instances of base::win::SamplingProfiler. A further CL will plumb this into a mechanism for polling running processes and attaching sampling profilers to modules of interest.

BUG=
R=sebmarchand@chromium.org, siggi@chromium.org

Review URL: https://codereview.appspot.com/10754043

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@1610 15e8cca8-e42c-11de-a347-f34a4f72eb7d