Move the logging code out from the HeapProxy.
The logging was shared between the HeapProxy and the AsanRuntime class, this design was really messy and impossible to maintain. Moving all the code to the AsanRuntime class allows to do a better/smarter logging (it's easier to put some stuff behind a flag). I've also enhanced the AsanErrorInfo struct and I use it directly to do the report, this way we get the same information if we're running under Windbg or a logger.

BUG=
R=chrisha@chromium.org

Review URL: https://codereview.appspot.com/11141043

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@1645 15e8cca8-e42c-11de-a347-f34a4f72eb7d