Add block attributes for special COFF blocks
These are: headers, symbol, string and relocation tables. The new
attributes are used in place of the old PE_PARSED attribute. This is to
allow easy and reliable identification of these special blocks, that
need separate treatment, in anticipation of COFF file writing support.

R=chrisha@chromium.org
BUG=

Review URL: https://codereview.appspot.com/12951043

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@1731 15e8cca8-e42c-11de-a347-f34a4f72eb7d