Adjust CodeView entries of PdbImageLayoutBuilderTest tests to prevent crashes under debugger.
It turns out that the debugger would sometimes crash while loading transformed test DLL images, if the transformed image was left with references to the original symbol file. Breaking the link to the symbol file prevents this from happening. This was only observed under Visual Studio, and not under cdb or windbg.

BUG=
R=rogerm@chromium.org, sebmarchand@chromium.org

Review URL: https://codereview.appspot.com/13301043

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@1756 15e8cca8-e42c-11de-a347-f34a4f72eb7d