Refactoring of basic block agent.
On CPU2006 Benchmark, the previous approach give a
17.44 slowdown.
The unbufferred version give 28.1X and the bufferred version give 18.30X.

This approach give about the same performance but avoid using a segment by thread, which was an issue when instrumenting chrome.dll.

A fastpath implementation is coming soon.

R=chrisha@chromium.org, sebmarchand@chromium.org
BUG=

Review URL: https://codereview.appspot.com/13471044

git-svn-id: http://sawbuck.googlecode.com/svn/trunk@1782 15e8cca8-e42c-11de-a347-f34a4f72eb7d